+++ mlprogram/samplers/__init__.py
@@ -1,5 +1,5 @@
 from mlprogram.samplers.sampler import SamplerState, DuplicatedSamplerState, Sampler, transform  # noqa
 from mlprogram.samplers.action_sequence_sampler import ActionSequenceSampler  # noqa
 from mlprogram.samplers.sampler_with_value_network import SamplerWithValueNetwork  # noqa
-from mlprogram.samplers.ast_reference_sampler import Reference, AstReferenceSampler  # noqa
+from mlprogram.samplers.sequential_program_sampler import Reference, SequentialProgramSampler  # noqa
 from mlprogram.samplers.filtered_sampler import FilteredSampler  # noqa
