+++ test/test_treegen.py
@@ -16,7 +16,7 @@ import mlprogram.nn
 from mlprogram.entrypoint import evaluate as eval, train_supervised
 from mlprogram.entrypoint import EvaluateSynthesizer
 from mlprogram.entrypoint.train import Epoch
-from mlprogram.entrypoint.torch import Optimizer
+from mlprogram.entrypoint.modules.torch import Optimizer
 from mlprogram.actions import AstToActionSequence
 from mlprogram.utils import Query, Token
 from mlprogram.synthesizers import BeamSearch
@@ -34,11 +34,11 @@ from mlprogram.nn.action_sequence import Loss, Predictor
 from mlprogram.nn import treegen
 from mlprogram.metrics import Accuracy
 
-from test.nl2code_dummy_dataset import is_subtype
-from test.nl2code_dummy_dataset import train_dataset
-from test.nl2code_dummy_dataset import test_dataset
-from test.nl2code_dummy_dataset import get_token_type
-from test.test_case_utils import integration_test
+from nl2code_dummy_dataset import is_subtype
+from nl2code_dummy_dataset import train_dataset
+from nl2code_dummy_dataset import test_dataset
+from nl2code_dummy_dataset import get_token_type
+from test_case_utils import integration_test
 
 logging.basicConfig(level=logging.INFO, stream=sys.stdout, force=True)
